feat: complete day 3 part 1
I feel as though this code is incredibly messy... It gets the job done,
but I'm not a huge fan of how I fixed it. Something tells me there's a
blatantly obvious way to solve this problem that I'm completely missing.

I would also like to remove the `expect()` calls at some point as well.
They feel like I'm cutting corners and shouldn't be in the final code.
The code repetition around those calls is driving me crazy as well.